requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
google-genai
beautifulsoup4
scikit-learn
numpy
requests
ultralytics
bytetracker
easyocr
openai-whisper
librosa
transformers
torch
torchvision
fpdf2
